fix(platform): move ExternalApplicationService out of common to break import cycle
from uipath.platform.identity import IdentityService fails in 0.1.89/0.1.90
with a circular import: identity imports common primitives, and
common/__init__.py eagerly imports _external_application_service, which
imports back into the still-initializing identity package. The cycle was
masked until #1781 made platform/__init__.py lazy.

Relocate ExternalApplicationService to a new uipath.platform.external_applications
domain package so dependencies flow one way: external_applications -> identity ->
common. No compat re-export in common (would recreate the cycle); org-wide search
found no consumers outside this monorepo. Add a subprocess-based import-order
regression test covering all entry points. Bump uipath-platform to 0.1.91 and
uipath to 2.12.7.
