+++ packages/uipath-platform/src/uipath/platform/external_applications/__init__.py
@@ -1,0 +1,5 @@
+"""UiPath External Applications."""
+
+from ._external_application_service import ExternalApplicationService
+
+__all__ = ["ExternalApplicationService"]
