+++ packages/uipath-platform/src/uipath/platform/external_applications/_external_application_service.py
@@ -7,9 +7,9 @@ from httpx import HTTPStatusError
 
 from uipath.platform.constants import ENV_BASE_URL
 
+from ..common.auth import TokenData
 from ..errors import EnrichedException
 from ..identity import IdentityService
-from .auth import TokenData
 
 
 class ExternalApplicationService:
